Adding mocking tests to #23 CLI
