+++ tests/test_knitlib.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 from click.testing import CliRunner
-import knitlib
-from knitlib.__main__ import main
+import mock
+import knitlib.plugins.dummy_plugin
+from knitlib import __main__ as main
 
 
 
@@ -10,28 +11,14 @@ def test_main():
     result = runner.invoke(main, [])
 
     assert result.output is not '()\n'
-    assert result.exit_code == 0
-
-def test_machine_handler_get_machine_by_id():
-    machine = knitlib.machine_handler.get_machine_plugin_by_id("DummyKnittingPlugin")
-    assert machine.__name__ == "DummyKnittingPlugin"
-
-def test_dummy_machine():
-
-    mach_type = knitlib.machine_handler.get_machine_types().other
-    other_type_dict = knitlib.machine_handler.get_machines_by_type(mach_type)
-    assert type(other_type_dict) is dict
-    dummy_type = other_type_dict["DummyKnittingPlugin"]()
-
-    dummy_type.configure(None)
-    dummy_type.knit()
-    dummy_type.finish()
+    assert result.exit_code is not 0  # It should fail due to invalid configuration,
 
 
-def test_ayab_plugin():
-    machine = knitlib.machine_handler.get_machine_plugin_by_id("AyabPluginControl")()
+@mock.patch('knitlib.plugins.dummy_plugin.logging')
+def test_dummy_knit(mock_logging):
+    runner = CliRunner()
+    result = runner.invoke(main, ["--plugin_name","DummyKnittingPlugin"])
 
-    #dummy_type.configure(None)
-    #dummy_type.knit() # https://bitbucket.org/chris007de/ayab-apparat/wiki/english/Hardware#!nomachine-development-mode
-    #if dummy_type.isstate("finished"):
-    #    dummy_type.finish()+
+    assert mock_logging.debug.assert_called()
+
+
