+++ tox.ini
@@ -26,6 +26,7 @@ deps =
     pytest
     pytest-capturelog
     pytest-cov
+    mock
 commands =
     python setup.py clean --all build_ext --force --inplace
     {posargs:py.test --cov=src --cov-report=term-missing -vv}
